Table Generator
Added a pair of functions to generate tables of values from any calculation function. Super helpful for a lot of things, and has a flag for formatting for the wiki.

Also added a function to calculate your paragon production bonus percentage, based on an amount of paragon, burned paragon, and how many sephirots you have.