Fixed AutoShatter
autoShatter now works.

Also, added goals for cycles. By default, all of them are 1, meaning they're allowed to be skipped. If autoshatter ends up in a cycle that has a value of 0 for the goal, it will stop skipping until that cycle is passed. You can use this to skip cycles you don't like and stay in ones you do.