Catnip calculator sanity check
Added final sanity check to catnip calculator. If theoretical and actual numbers match, then it'll return only the actuals for simplicity. If they don't, it'll return both for further checking.

Note: The numbers will *not* match up if you pass in alternate quantities in the parameters. In that case, so long as the theoretical is accurate (check with normal numbers to see), then only check the theoretical.